requirement.txt
blinker==1.5
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous @ git+https://github.com/pallets/itsdangerous.git@de09cad488a4d7c7bbcbcdb8e1c2dfde64325f48
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector==2.2.9
PyJWT==2.6.0
SQLAlchemy==2.0.1
typing_extensions==4.4.0
Werkzeug==2.2.2
WTForms==3.0.1
